Remove sleep
