+++ crates/full-node/sov-sequencer/tests/integration/utils.rs
@@ -335,18 +335,3 @@ pub fn tx_set_value_with_gas<RT: Runtime<TestSpec> + EncodeCall<ValueSetter<Test
 pub fn tempdir_inside_codebase_dir() -> Arc<tempfile::TempDir> {
     Arc::new(tempfile::tempdir_in(std::env!("CARGO_TARGET_TMPDIR")).unwrap())
 }
-
-use serde::Deserialize;
-use sov_node_client::NodeClient;
-
-pub async fn get_height(client: &NodeClient) -> anyhow::Result<RollupHeight> {
-    #[derive(Deserialize, Debug)]
-    struct Data {
-        value: (u64, u64),
-    }
-
-    let url = "/modules/chain-state/state/current-heights";
-    let response = client.http_get(url).await?;
-    let heights: Data = serde_json::from_str(&response)?;
-    Ok(RollupHeight::new(heights.value.0))
-}
